+++ hss_okta_flutter/lib/src/web/hss_okta_js.dart
@@ -10,6 +10,26 @@ class OktaAuth {
   external factory OktaAuth(OktaConfig options);
   external Token get token;
   external TokenManager get tokenManager;
+
+  /// Parses tokens from the redirect url and stores them.
+  external Future<void> storeTokensFromRedirect();
+
+  ///Handle a redirect to the configured redirectUri that happens on the end of login flow, enroll authenticator flow or on an error.
+  ///Stores tokens from redirect url into storage (for login flow), then redirect users back to the originalUri.
+  /// When using PKCE authorization code flow, this method also exchanges authorization code for tokens.
+  /// By default it calls window.location.replace for the redirection.
+  /// The default behavior can be overrided by providing options.restoreOriginalUri.
+  /// By default, originalUri will be retrieved from storage, but this can be overridden by specifying originalUri in the first parameter to this function.
+  external Future<void> handleRedirect(String? originalUri);
+
+  /// Check window.location to verify if the app is in OAuth callback state or not. This function is synchronous and returns true or false.
+  external bool isRedirect();
+
+  /// Returns the access token string retrieved from [AuthState] if it exists.
+  external String getAccessToken();
+
+  ///Returns the id token string retrieved from [AuthState] if it exists.
+  external String getIdToken();
 }
 
 /// Used With [OktaAuth]'s Initializer
@@ -22,9 +42,17 @@ class OktaConfig {
     String? redirectUri,
   });
 
-  external String issuer;
-  external String clientId;
-  external String redirectUri;
+  external String? issuer;
+  external String? clientId;
+  external String? redirectUri;
+  external String? scopes;
+  external String? state;
+  external bool? pkce;
+  external String? authorizeUrl;
+  external String? userinfoUrl;
+  external String? tokenUrl;
+  external String? revokeUrl;
+  external String? logoutUrl;
 }
 
 /// Contains methods for accessing tokens.
@@ -145,4 +173,18 @@ class IDToken {
   external String idToken;
   external String issuer;
   external String clientId;
+}
+
+///AuthStateManager evaluates and emits AuthState based on the events from TokenManager for downstream clients to consume.
+@JS()
+class AuthStateManager {
+  external Future<AuthState?> getAuthState();
+}
+
+@JS()
+class AuthState {
+  external bool isAuthenticated;
+  external String accessToken;
+  external String idToken;
+  external String error;
 }